+++ ci-internal/JenkinsfilePullRequestJob.groovy
@@ -156,6 +156,7 @@ pipeline.stages = [
         pipeline.stage(PRE_MERGE) {
             CommonUtil.safe(script) {
                 script.sh "git reset --merge" //revert previous failed merge
+                script.sh "git clean -fd"
             }
             script.git(
                     url: pipeline.repoUrl,
@@ -239,7 +240,11 @@ pipeline.stages = [
             script.sh("echo \"androidStandardDebugDir=$workspace\n" +
                     "androidStandardDebugMode=true\n" +
                     "skipSamplesBuild=true\" > template/android-standard/androidStandard.properties")
-            script.sh("./gradlew -p template clean build assembleQa --stacktrace")
+            /**
+             * assembleDebug is used for assembleAndroidTest with testBuildType=debug for Template.
+             * Running assembleAndroidTest with testBuildType=qa could cause some problems with proguard settings
+             */
+            script.sh("./gradlew -p template clean build assembleQa assembleDebug --stacktrace")
         },
         pipeline.stage(UNIT_TEST) {
             AndroidPipelineHelper.unitTestStageBodyAndroid(script,